requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
psutil>=5.9.0
aiofiles>=23.2.1
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
